refactor: 컴포넌트 파일에서 컴포넌트가 아닌 export 를 분리
- 훅·상수를 함께 내보내면 개발 중 파일 수정 시 화면이 통째로 새로고침된다
