+++ frontend/src/components/ui/badge.tsx
@@ -45,4 +45,4 @@ function Badge({
   )
 }
 
-export { Badge, badgeVariants }
+export { Badge }
