+++ frontend/src/components/ui/button.tsx
@@ -61,4 +61,4 @@ function Button({
   )
 }
 
-export { Button, buttonVariants }
+export { Button }
