+++ frontend/src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
-import { AuthProvider } from './contexts/AuthContext'
-import { RoundProvider } from './contexts/RoundContext'
+import { AuthProvider } from './contexts/AuthProvider'
+import { RoundProvider } from './contexts/RoundProvider'
 import './index.css'
 import App from './App.tsx'
